requirements.txt
Flask
mysql-connector-python
flask-hcaptcha
gunicorn
